requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.0
django-phonenumber-field==5.0.0
gunicorn==20.0.4
phonenumbers==8.12.9
Pillow==2.2.1
pkg-resources==0.0.0
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
whitenoise==5.2.0
